JS root tests do not interfre with other tests now.
