+++ src/Components/test/E2ETest/Tests/StatePersistenceTest.cs
@@ -355,20 +355,6 @@ public class StatePersistenceTest : ServerTestBase<BasicTestAppServerSiteFixture
             interactiveRuntime: interactiveRuntime);
     }
 
-    [Theory]
-    [InlineData("ServerNonPrerendered")]
-    [InlineData("WebAssemblyNonPrerendered")]
-    public void PersistentStateIsSupportedInDynamicJSRoots(string renderMode)
-    {
-        Navigate($"subdir/WasmMinimal/dynamic-js-root.html?renderMode={renderMode}");
-
-        Browser.Equal("Counter", () => Browser.Exists(By.TagName("h1")).Text);
-        Browser.Equal("Current count: 0", () => Browser.Exists(By.CssSelector("p[role='status']")).Text);
-
-        Browser.Click(By.CssSelector("button.btn-primary"));
-        Browser.Equal("Current count: 1", () => Browser.Exists(By.CssSelector("p[role='status']")).Text);
-    }
-
     private void AssertPageState(
         string mode,
         string renderMode,
