+++ src/Components/test/testassets/Components.TestServer/RazorComponentEndpointsStartup.cs
@@ -51,7 +51,10 @@ public class RazorComponentEndpointsStartup<TRootComponent>
                     options.DisconnectedCircuitMaxRetained = 0;
                     options.DetailedErrors = true;
                 }
-                options.RootComponents.RegisterForJavaScript<TestContentPackage.PersistentComponents.ComponentWithPersistentState>("dynamic-js-root-counter");
+                if (Configuration.GetValue<bool>("RegisterDynamicJSRootComponent"))
+                {
+                    options.RootComponents.RegisterForJavaScript<TestContentPackage.PersistentComponents.ComponentWithPersistentState>("dynamic-js-root-counter");
+                }
             })
             .AddAuthenticationStateSerialization(options =>
             {
